# docs/swagger.yaml
basePath: /api
definitions:
...
  ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/HttpError'
    type: object
...
  Friend:
    properties:
      id:
        type: string
      image:
        type: string
      isOnline:
        type: boolean
      username:
        type: string
    type: object
  FriendRequest:
    properties:
      id:
        type: string
      image:
        type: string
      type:
        description: '1: Incoming, 0: Outgoing'
        enum:
        - 0
        - 1
        type: integer
      username:
        type: string
    type: object
...
  HttpError:
    properties:
      message:
        description: The specific error message
        type: string
      type:
        description: The Http Response as a string
        type: string
    type: object
...
  SuccessResponse:
    properties:
      success:
        description: Only returns true, not a json object
        type: boolean
    type: object
...
host: localhost:<PORT>
info:
  contact: {}
  description: Xonia REST API . This service uses sessions for authentication
  license:
    name: GNUv3
  title: Xonia API
  version: "2.0.0"
paths:
...
  /account/{memberId}/friend:
    delete:
      parameters:
      - description: User ID
        in: path
        name: memberId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Remove Friend
      tags:
      - Friends
    post:
      parameters:
      - description: User ID
        in: path
        name: memberId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Send Friend Request
      tags:
      - Friends
  /account/{memberId}/friend/accept:
    post:
      parameters:
      - description: User ID
        in: path
        name: memberId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Accept Friend's Request
      tags:
      - Friends
  /account/{memberId}/friend/cancel:
    post:
      parameters:
      - description: User ID
        in: path
        name: memberId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Cancel Friend's Request
      tags:
      - Friends
...
  /account/me/friends:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Friend'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get Current User's Friends
      tags:
      - Friends
  /account/me/pending:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/FriendRequest'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get Current User's Friend Requests
      tags:
      - Friends
...
swagger: "2.0"